Revert "test(logging): add profiling test (#2000)" (#2006)
